+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/registry/HandlerAdapterRegistry.java
@@ -1,4 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.adapter;
+package com.interface21.webmvc.servlet.mvc.registry;
+
+import com.interface21.webmvc.servlet.mvc.adapter.HandlerAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
