+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/registry/HandlerMappingRegistry.java
@@ -1,5 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.mapping;
+package com.interface21.webmvc.servlet.mvc.registry;
 
+import com.interface21.webmvc.servlet.mvc.mapping.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 
 import java.util.ArrayList;
