+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/servlet/DispatcherServlet.java
@@ -1,13 +1,14 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.servlet;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
 import com.interface21.webmvc.servlet.mvc.adapter.AnnotatedHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.adapter.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.adapter.HandlerAdapterRegistry;
+import com.interface21.webmvc.servlet.mvc.registry.HandlerAdapterRegistry;
 import com.interface21.webmvc.servlet.mvc.adapter.ManualHandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.mapping.HandlerMappingRegistry;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
+import com.interface21.webmvc.servlet.mvc.registry.HandlerMappingRegistry;
+import com.interface21.webmvc.servlet.mvc.mapping.AnnotationHandlerMapping;
+import com.interface21.webmvc.servlet.mvc.mapping.ManualHandlerMapping;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
